Crawly pulls page
I can pull a page down with crawly but I'm running into trouble with navigating to the correct element. I can get as far down as the "nethys-search" element but it's contents are not populated. I'll need to either find a way to trigger the mechanism that populates that element or look into an alternative approach.